Make net(...) smart about local host & network availability
Change `net(...)` to not test if a network is available if the given
destination address is on the local host. This prevents incorrectly
returning `NetworkFailure` when the current host is detached from the
network and the failure is actually a `ServiceFailure` (for example,
if nothing is listening on the destination port on the local host).